requirements.txt
Flask==3.0.3
twilio==9.3.3
python-dotenv==1.0.1
pymongo
openai>=1.0.0
